requirements.txt
torch==2.3.1
torchmetrics=1.4.0
torch_geometric==2.5.3
comet_ml==3.43.0
lightning==2.2.5
hydra-core==1.3.2